+++ src/main/java/library/commandrunner/loan/LoanCommandRunner.java
@@ -1,5 +1,6 @@
-package library.commands;
+package library.commandrunner.loan;
 
+import library.commandrunner.LibraryCommandRunner;
 import library.customer.Customer;
 import library.item.Borrowable;
 import library.item.Item;
